+++ interface/interface.ts
@@ -1,0 +1,6 @@
+export interface item {
+  index: number;
+  title: string;
+};
+
+export interface items extends Array<item> {}
