+++ package.json
@@ -19,7 +19,7 @@
     "autoprefixer": "10.4.14",
     "eslint": "8.44.0",
     "eslint-config-next": "13.4.8",
-    "next": "13.4.8",
+    "next": "^13.4.9",
     "postcss": "8.4.24",
     "react": "18.2.0",
     "react-dom": "18.2.0",
